Added JWT Authentication and Authorization

+++ WishList/WishList.App/Controller/WishesController.cs
@@ -1,11 +1,14 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using WishList.BusinessLogic.Models;
+using WishList.Services.Exceptions;
 using WishList.Services.Interfaces;
 
 namespace WishList.App.Controller
 {
     [Route("[controller]")]
     [ApiController]
+    [Authorize]
     public class WishesController : ControllerBase
     {
         private readonly IWishService wishService;
@@ -18,7 +21,8 @@ namespace WishList.App.Controller
         [HttpPost]
         public async Task<ActionResult> CreateWishAsync(CreateWishDto wish)
         {
-            await wishService.AddWishAsync(wish.UserId, wish);
+            var currentUser = (Infrastructure.Models.User)HttpContext.Items["User"];
+            await wishService.AddWishAsync(currentUser, wish);
             return Ok();
         }
 
@@ -37,6 +41,13 @@ namespace WishList.App.Controller
         [HttpDelete]
         public async Task<ActionResult> DeleteWish(int wishId)
         {
+            var wish = await wishService.GetWishById(wishId);
+            var currentUser = (Infrastructure.Models.User)HttpContext.Items["User"];
+            if (wish.UserId != currentUser.Id)
+            {
+                throw new AppException() { StatusCode = System.Net.HttpStatusCode.BadRequest, Message = "It is a wish of another user." };
+            }
+
             return Ok(await wishService.DeleteWishAsync(wishId));
         }
     }
